+++ src/lightning/fabric/plugins/precision/double.py
@@ -30,7 +30,7 @@ class DoublePrecision(Precision):
     def convert_module(self, module: Module) -> Module:
         return module.double()
 
-    def init_context(self) -> ContextManager:
+    def module_init_context(self) -> ContextManager:
         return _DtypeContextManager(torch.double)
 
     def forward_context(self) -> ContextManager: