+++ src/lightning/pytorch/plugins/precision/half.py
@@ -41,7 +41,7 @@ class HalfPrecisionPlugin(PrecisionPlugin):
         return module.to(dtype=self._desired_input_dtype)
 
     @contextmanager
-    def init_context(self) -> Generator[None, None, None]:
+    def module_init_context(self) -> Generator[None, None, None]:
         """A context manager to change the default tensor type when initializing module parameters or tensors.
 
         See: :func:`torch.set_default_dtype`